+++ BankStartWeb/Pages/Customer/CustomerDetail.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model BankStartWeb.Pages.CustomerDetailModel
+@model BankStartWeb.Pages.Customer.CustomerDetailModel
 @{
 }
