--- BankStartWeb/Pages/Customer/EditCustomer.cshtml
@@ -1,4 +1,0 @@
-﻿@page
-@model BankStartWeb.Pages.Customer.EditCustomerModel
-@{
-}
