+++ BankStartWeb/Pages/Customer/EditCustomer2.cshtml
@@ -1,0 +1,4 @@
+﻿@page
+@model BankStartWeb.Pages.Customer.EditCustomer2Model
+@{
+}
